requirements.txt
python-telegram-bot[ext]
firebase-admin
openai
python-dotenv
